Revert "minor refactor"
This reverts commit 6ab8bfb3ece130c84f0a3de451dc247a9fed79db.
